requirements.txt
asgiref==3.6.0
attrs==23.1.0
beautifulsoup4==4.12.2
bootstrap4==0.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
crispy-bootstrap4==2022.1
cssselect2==0.7.0
Django==4.2
django-bootstrap4==23.1
django-crispy-forms==2.0
django-pdfkit==0.3.1
django-widget-tweaks==1.4.12
exceptiongroup==1.1.2
fonttools==4.41.0
h11==0.14.0
html5lib==1.1
idna==3.4
outcome==1.2.0
pdfkit==0.6.0
Pillow==10.0.0
pycparser==2.21
pydyf==0.7.0
pyphen==0.14.0
PySocks==1.7.1
python-dateutil==2.8.2
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
weasyprint==59.0
webencodings==0.5.1
wkhtmltopdf==0.2
wsproto==1.2.0
zopfli==0.2.2
